refactor(admin): split PluginsTab + ServicesTab into hooks + presentational (#138)
Two monolithic admin tabs (768 + 357 lines) decomposed into their data
layer, presentational components, and thin shells. Zero UX regression —
same JSX, same backend calls, same modals.

ServicesTab (357 → 105 lines):
  services/
    useServicesTab.ts  — fetch / test / toggle / setDefault / delete
    ServiceRow.tsx      — row card with action cluster
    ServiceModal.tsx    — create/edit form + Plex helpers

PluginsTab (768 → 198 lines):
  plugins/
    constants.ts          — RegistryPlugin, CATEGORY_CONFIG, TAG_CONFIG
    usePluginsTab.ts       — list + registry + mutations + restart poller
    PluginCardChrome.tsx   — CopyButton + PluginInitial helpers
    InstalledList.tsx      — installed grid + empty state
    DiscoverList.tsx       — discover grid + loading/error/empty states
    ManualInstallModal.tsx — advanced git-clone escape hatch
    PluginDocsModal.tsx    — plugin-dev quickref

Reviewer fixes applied before commit:
- install() now returns a boolean so the consent modal stays open on error
  (old code closed the modal + flipped sub-tab even on failure, hiding the
  error toast that only renders inside Discover)
- deleteService() same pattern — returns a boolean, consumer only closes
  the confirm modal on success
- fetchPlugins() in the success path is fire-and-forget so the consent
  modal unmounts immediately instead of waiting on the refetch round-trip
